Fix build for Windows MinGW GCC 6.4 [INFRA-151] (#129)
Allows libswiftnav to build on Windows MinGW GCC 6.4 by fixing warnings that prevented compilation.  MinGW allows us to build unix style tools on Windows easily with Travis, in particular gnss-converters for https://github.com/swift-nav/gnss-converters/pull/363.

Additional fixes:
+ Add Windows MinGW GCC 6.4 to the build matrix per above
+ Fix GCC 6.4 to use bionic since a gcc-6 package is available there
+ Make Travis fetch submodules, instead of each build script doing it manually

---

* Fix build for Windows MinGW GCC 6.4

* Clang format

* Use bionic for gcc-6

* Add windows mingw 6.4 build

* Remove sh.exe to fix cmake error

* Tell travis to get submodules

* Add fix-up for missing srandom/random

+++ src/gnss_time.c
@@ -29,7 +29,8 @@
  * \return true if the time is valid
  */
 bool gps_time_valid(const gps_time_t *t) {
-  return isfinite(t->tow) && t->tow >= 0 && t->tow < WEEK_SECS && t->wn >= 0;
+  return isfinite((float)t->tow) && t->tow >= 0 && t->tow < WEEK_SECS &&
+         t->wn >= 0;
 }
 
 /** Tell whether a `time_t` struct is a valid Unix time.
@@ -38,7 +39,7 @@ bool gps_time_valid(const gps_time_t *t) {
  * \return true if the time is valid
  */
 bool unix_time_valid(const time_t *t) {
-  return (isfinite((double)*t) && ((*t) >= 0) &&
+  return (isfinite((float)*t) && ((*t) >= 0) &&
           ((*t) > (GPS_WEEK_REFERENCE * WEEK_SECS) + GPS_EPOCH));
 }
